+++ kaithem/src/docs/changes.md
@@ -7,7 +7,9 @@ Change Log
 - :coffin: Remove old baresip code
 - :coffin: Remove kaithem.midi API
 - :coffin: Remove the image map creator util
-- :coffin: Remove kaithem.time.accuracy
+- :coffin: Remove kaithem.time.accuracy, lantime, hour, month, day, second, minute, dayofweek
+  - :coffin: Remove kaithem.sys.shellex, shellexbg, which
+
 - :coffin: Remove kaithem.events.when and kaithem.events.after
 
 - :sparkles: Split off the sound stuff in a separate libary [IceMedia](https://github.com/EternityForest/icemedia) meant for easy standalone use.